pyproject.toml
[tool.poetry]
name = "fastapi-blog"
version = "0.1.0"
description = ""
authors = ["Sergey Fomichev <fomichev.ser.v@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.11.1"
amqp = "5.1.1"
annotated-types = "0.5.0"
anyio = "3.7.1"
async-timeout = "4.0.2"
asyncpg = "0.28.0"
attrs = "23.1.0"
bcrypt = "4.0.1"
billiard = "4.1.0"
celery = "5.3.1"
certifi = "2023.5.7"
cffi = "1.15.1"
charset-normalizer = "3.2.0"
click = "8.1.4"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
crypto = "1.4.1"
cryptography = "41.0.2"
dnspython = "2.3.0"
ecdsa = "0.18.0"
email-validator = "2.0.0.post2"
envparse = "0.2.0"
exceptiongroup = "1.1.2"
fastapi = "0.100.0"
fastapi-cache2 = "0.2.1"
fastapi-pagination = "0.12.5"
flower = "2.0.0"
greenlet = "2.0.2"
h11 = "0.14.0"
httpcore = "0.17.3"
httpx = "0.24.1"
humanize = "4.7.0"
idna = "3.4"
isort = "5.12.0"
kombu = "5.3.1"
makefun = "1.15.1"
mako = "1.2.4"
markupsafe = "2.1.3"
naked = "0.1.32"
packaging = "23.1"
passlib = "1.7.4"
pendulum = "2.1.2"
prometheus-client = "0.17.1"
psycopg2-binary = "2.9.6"
pyasn1 = "0.5.0"
pycparser = "2.21"
pydantic = "2.0.3"
pydantic-core = "2.3.0"
pyjwt = "2.7.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-jose = "3.3.0"
python-multipart = "0.0.6"
python-slugify = "8.0.1"
pytz = "2023.3"
pytzdata = "2020.1"
pyyaml = "6.0"
redis = "4.6.0"
requests = "2.31.0"
rsa = "4.9"
shellescape = "3.8.1"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.18"
starlette = "0.27.0"
text-unidecode = "1.3"
tornado = "6.3.2"
typing-extensions = "4.7.1"
tzdata = "2023.3"
urllib3 = "2.0.3"
uvicorn = "0.22.0"
vine = "5.0.0"
gunicorn = "^22.0.0"

[tool.poetry.dev-dependencies]
flake8 = "6.0.0"
iniconfig = "2.0.0"
mccabe = "0.7.0"
pycodestyle = "2.10.0"
pyflakes = "3.0.1"
pytest = "7.2.0"
pytest-asyncio = "0.20.3"
tomli = "2.0.1"
shutup = "0.2.0"
prompt-toolkit = "3.0.39"
pluggy = "1.2.0"
